[ADD] abstract class for user resolve event
